Splittet ListevelgerEnkelt og ListevelgerGruppe.
