+++ src/felles-komponenter/skjema/index.js
@@ -3,7 +3,6 @@ import Input from './input/input';
 import Radio from './input/radio';
 import RadioGruppe from './input/radio-gruppe';
 import Select from './input/select';
-import FriSelect from './input/fri-select';
 import Textarea from './textarea/textarea';
 import Fieldset from '../../utils/navFrontend';
 import * as Validering from './validering';
@@ -16,7 +15,6 @@ export {
   Radio,
   RadioGruppe,
   Select,
-  FriSelect,
   Textarea,
   Validering,
   Fieldset,
